+++ src/components/home-button/HomeButton.jsx
@@ -1,15 +1,8 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import classnames from 'classnames';
-import {TYPE} from 'logo/Logo';
+import {TYPE, BASE_URL, LOGOS} from 'logo/Logo';
 
-const BASE_URL = 'https://styleguide.brainly.com/images/logos/';
-const LOGOS = {
-  'brainly': 'brainly-761d75d6ea',
-  'eodev': 'eodev-1972bd4349',
-  'nosdevoirs': 'nosdevoirs-e2d5d17215',
-  'znanija': 'znanija-addd85e6f5'
-};
 const ICONS = {
   'brainly': 'brainly-mobile-426ef8718f',
   'eodev': 'eodev-mobile-b3319881d5',
