+++ src/components/logo/Logo.jsx
@@ -2,11 +2,20 @@ import React from 'react';
 import classnames from 'classnames';
 import PropTypes from 'prop-types';
 
+const BASE_URL = 'https://styleguide.brainly.com/images/logos/';
+
 const TYPE = {
   BRAINLY: 'brainly',
   EODEV: 'eodev',
   NOSDEVOIRS: 'nosdevoirs',
   ZNANIJA: 'znanija'
+};
+
+const LOGOS = {
+  'brainly': 'brainly-761d75d6ea',
+  'eodev': 'eodev-1972bd4349',
+  'nosdevoirs': 'nosdevoirs-e2d5d17215',
+  'znanija': 'znanija-addd85e6f5'
 };
 
 const Logo = ({type = TYPE.BRAINLY, className}) => {
@@ -15,7 +24,7 @@ const Logo = ({type = TYPE.BRAINLY, className}) => {
     [`sg-logo--${type}`]: type !== TYPE.BRAINLY
   }, className);
 
-  const logoPath = `images/logos/${type}.svg`;
+  const logoPath = `${BASE_URL}${LOGOS[type]}.svg`;
 
   return <div className={logoClass}>
     <img className="sg-logo__image" src={logoPath}/>
@@ -28,4 +37,4 @@ Logo.propTypes = {
 };
 
 export default Logo;
-export {TYPE};
+export {TYPE, BASE_URL, LOGOS};
